feat: use dialog with backdrop to avoid weird interactions behind the form
* refactor: move `AtomLabelEditForm` in his own file

Closes: https://github.com/zakodium-oss/react-ocl/issues/80
